TUC-129 More cleaning thanks to clang-tidy
	modified:   checkcode.sh
	modified:   ../src/tucucli/tucucli.cpp
	modified:   ../src/tucucommon/datetime.cpp
	modified:   ../src/tucucommon/licensemanager.cpp
	modified:   ../src/tucucommon/logger.cpp
	modified:   ../src/tucucommon/systeminfo.cpp
	modified:   ../src/tucucommon/translatablestring.cpp
	modified:   ../src/tucucommon/utils.cpp
	modified:   ../src/tucucommon/xmldocument.cpp
	modified:   ../src/tucucore/cachedexponentials.cpp
	modified:   ../src/tucucore/computingcomponent.cpp
	modified:   ../src/tucucore/computingcomponent.h
	modified:   ../src/tucucore/computingresponseexport.cpp
	modified:   ../src/tucucore/concentrationcalculator.cpp
	modified:   ../src/tucucore/concentrationcalculator.h
	modified:   ../src/tucucore/concentrationprediction.h
	modified:   ../src/tucucore/covariateextractor.cpp
	modified:   ../src/tucucore/covariateextractor.h
	modified:   ../src/tucucore/cyclestatisticscalculator.cpp
	modified:   ../src/tucucore/dosage.cpp
	modified:   ../src/tucucore/dosage.h
	modified:   ../src/tucucore/drugdomainconstraintsevaluator.cpp
	modified:   ../src/tucucore/drugfilevalidator.cpp
	modified:   ../src/tucucore/drugmodelchecker.cpp
	modified:   ../src/tucucore/drugmodelimport.cpp
	modified:   ../src/tucucore/drugmodelrepository.cpp
	modified:   ../src/tucucore/hardcodedoperation.cpp
	modified:   ../src/tucucore/intakeextractor.cpp
	modified:   ../src/tucucore/intakeintervalcalculator.cpp
	modified:   ../src/tucucore/likelihood.cpp
	modified:   ../src/tucucore/montecarlopercentilecalculator.cpp
	modified:   ../src/tucucore/operation.cpp
	modified:   ../src/tucucore/overloadevaluator.cpp
	modified:   ../src/tucucore/parameterids.h
	modified:   ../src/tucucore/percentilesprediction.h
	modified:   ../src/tucucore/version.cpp
	modified:   ../src/tuculicense/main.cpp
	modified:   ../src/tucuquery/parametersdata.h
	modified:   ../src/tucuquery/queryimport.cpp
	modified:   ../src/tucuquery/queryimport.h
	modified:   ../src/tucuvalidator/tucuvalidator.cpp
	modified:   ../test/tucucore/test_covariateextractor.h

+++ src/tucucommon/utils.cpp
@@ -153,14 +153,14 @@ std::string Utils::varToString(const DateTime &_value)
     char buf[64];
     sprintf(buf, "%04d-%02d-%02d %02d:%02d:%02d",
             _value.year(), _value.month(), _value.day(),
-            _value.hour(), _value.minute(), (int)_value.second());
+            _value.hour(), _value.minute(), _value.second());
     return std::string(buf);
 }
 
 
 DateTime Utils::ValueToDate(const Value &_value)
 {
-    DateTime dt(Duration(std::chrono::seconds((int64)_value)));
+    DateTime dt(Duration(std::chrono::seconds(static_cast<int64>(_value))));
 
     return dt;
 }
